requirements.txt
docx==0.2.4
matplotlib==3.9.0
numpy==1.24.3
pandas==2.2.2
Pillow==10.3.0
plotly==5.22.0
PyPDF2==3.0.1
python_docx==0.8.11
streamlit==1.35.0
wordcloud==1.9.3
